Time: 28 ms (55.16%), Space: 88.3 MB (7.85%) - LeetHub

+++ 2226-maximum-candies-allocated-to-k-children/2226-maximum-candies-allocated-to-k-children.cpp
@@ -1,6 +1,6 @@
 class Solution {
 private:
-    bool isOK(long long maximumCandy, vector<int>& candies, long long k) {
+    bool isOK(int maximumCandy, vector<int>& candies, long long k) {
         long long cnt = 0;
         for (int c : candies) {
             cnt += c / maximumCandy;
@@ -9,9 +9,9 @@ private:
     }
 public:
     int maximumCandies(vector<int>& candies, long long k) {
-        long long left = 1, right = *max_element(candies.begin(), candies.end());
+        int left = 1, right = *max_element(candies.begin(), candies.end());
         while (left <= right) {
-            long long mid = left + (right - left) / 2;
+            int mid = left + (right - left) / 2;
             if (isOK(mid, candies, k)) {
                 left = mid + 1;
             }
